backend/requirements.txt
numpy
matplotlib
fastapi
uvicorn
pydantic
python-dotenv
